Function rle_decode_fast::append_from_within

fn append_from_within<T>(seif: &mut Vec<T>, src: Range<usize>) where
    T: Copy, 

Copy of vec::append_from_within() proposed for inclusion in stdlib, see https://github.com/rust-lang/rfcs/pull/2714 Heavily based on the implementation of slice::copy_within(), so we're pretty sure the implementation is sound

Note that the generic bounds were replaced by an explicit a..b range. This is so that we can compile this on older toolchains (< 1.28).