Crate rle_decode_fast

rle-decode-helper

THE fastest way to implement any kind of decoding for Run Length Encoded data in Rust.

Writing a fast decoder that is also safe can be quite challenging, so this crate is here to save you the hassle of maintaining and testing your own implementation.

Usage

let mut decode_buffer = vec![0, 0, 1, 1, 0, 2, 3];
let lookbehind_length = 4;
let output_length = 10;
rle_decode_fast::rle_decode(&mut decode_buffer, lookbehind_length, output_length);
assert_eq!(decode_buffer, [0, 0, 1, 1, 0, 2, 3, 1, 0, 2, 3, 1, 0, 2, 3, 1, 0]);

Functions

append_from_within

Copy of vec::append_from_within() proposed for inclusion in stdlib, see https://github.com/rust-lang/rfcs/pull/2714 Heavily based on the implementation of slice::copy_within(), so we’re pretty sure the implementation is sound

lookbehind_length_fail

rle_decode

Fast decoding of run length encoded data